feat: add a docs conversion tool for references
Takes the output of talosctl docs and converts it into .mdx formatted
files. Adds new make target to generate the docs based on which version
of the docs you want to generate.

Requires the generated files to still be added to yaml config files and
docs.json to be generated

Signed-off-by: Justin Garrison <justin.garrison@siderolabs.com>